server/requirements.txt
langchain
langgraph
langchain_community
langchain_groq
langchain_core
langchain-cli[inmem]
fastapi
uvicorn[standard]
python-dotenv
arxiv
wikipedia
requests
langchain-tavily
tavily-python